+++ app/helpers/admin/policy_configurations_helper.rb
@@ -1,8 +1,7 @@
 module Admin
   module PolicyConfigurationsHelper
     def options_for_academic_year
-      years = 4
-      (0..(years - 1)).map { |relative_year| AcademicYear.current + relative_year }
+      (0..3).map { |relative_year| AcademicYear.current + relative_year }
     end
   end
 end