+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog]
   - Add ECP admin claim identity confirmation task screen
   - Create admin claim note when identity confirmation mailer is sent
   - Amend admin timeout from 30 mins to 60 mins
+  - Add ECP admin claim qualifications task screen
 - All policies
   - Home Address selection
     - Add home address search screen to claim journey